move login/logout/register to pages
