+++ src/templates/pages/register.html
